refactoring request filter
