+++ WhataCookUsers/src/main/java/com/whatacook/cookers/config/filter/EmailResendFlowHandler.java
@@ -1,0 +1,9 @@
+package com.whatacook.cookers.config.filter;
+
+import org.springframework.web.server.ServerWebExchange;
+import org.springframework.web.server.WebFilterChain;
+import reactor.core.publisher.Mono;
+
+public interface EmailResendFlowHandler {
+    Mono<Void> handle(String keyEmailToResend, ServerWebExchange exchange, WebFilterChain chain);
+}
