+++ WhataCookUsers/src/main/java/com/whatacook/cookers/config/filter/EmailResetPasswordFlowHandler.java
@@ -1,0 +1,9 @@
+package com.whatacook.cookers.config.filter;
+
+import org.springframework.web.server.ServerWebExchange;
+import org.springframework.web.server.WebFilterChain;
+import reactor.core.publisher.Mono;
+
+public interface EmailResetPasswordFlowHandler {
+    Mono<Void> handle(String keyResetCode, ServerWebExchange exchange, WebFilterChain chain);
+}
