+++ WhataCookUsers/src/main/java/com/whatacook/cookers/config/filter/ResponseErrorHtml.java
@@ -1,0 +1,8 @@
+package com.whatacook.cookers.config.filter;
+
+import org.springframework.web.server.ServerWebExchange;
+import reactor.core.publisher.Mono;
+
+public interface ResponseErrorHtml {
+    Mono<Void> send(ServerWebExchange exchange, String htmlContent);
+}
