+++ WhataCookUsers/src/main/java/com/whatacook/cookers/config/filter/SetNewPasswordFlowHandler.java
@@ -1,0 +1,9 @@
+package com.whatacook.cookers.config.filter;
+
+import org.springframework.web.server.ServerWebExchange;
+import org.springframework.web.server.WebFilterChain;
+import reactor.core.publisher.Mono;
+
+public interface SetNewPasswordFlowHandler {
+    Mono<Void> handle(String keyCodeToSet, ServerWebExchange exchange, WebFilterChain chain);
+}
